fix: quickjs-wasm-sys: properly handle shifts by 32 in libbf.c
As reported upstream in https://github.com/bellard/quickjs/issues/131:
shifting by `(LIMB_BITS - shift)` when `LIMB_BITS == 32` and `shift == 0`
produces wrong results on 32-bit targets.

This fix replaces all occurences of `(LIMB_BITS - shift)` with
calls to the existing helpers: `shld()` and `shrd`, which are
moved to the beginning of the file so that they are defined
before their new usages.

The helpers properly handle the cases when `shift` is zero.

Fixes issue 556